Socket and HTTP Response bug fixes
fr::TcpSocket checked if the recv response was greater than 0, which created issues as 0 is a valid result if no data was received and no error occurred.

fr::HttpResponse::construct() now adds in a content-length header.

fr::HttpResponse::parse trims the body if it's larger than content length.

fr::Http::set_uri() now prepends a '/' if one's not provided.

Fixed fr::Http::parse_header from cutting off the end of headers.

+++ src/HttpRequest.cpp
@@ -53,13 +53,13 @@ namespace fr
         return true;
     }
 
-    bool HttpRequest::parse_header(int32_t header_end_pos)
+    bool HttpRequest::parse_header(int64_t header_end_pos)
     {
         try
         {
             //Split the header into lines
             size_t line = 0;
-            std::vector<std::string> header_lines = split_string(body.substr(0, header_end_pos));
+            std::vector<std::string> header_lines = split_string(body.substr(0, (unsigned long)header_end_pos));
             if(header_lines.empty())
                 return false;
